New version of "Parallel-executing" support Added:  -Added ability for changing name of "_connection" magic variable  -Added support for old Gevent
Fixed:
 -"VirtVar" class moved to separate module
 -Fixed default settings for manual GC
 -Fixed default settings for "antifreeze" system
 -Fixed calculating of "processingRequestCount"
 -Message about experimental futures when it not used
 -Copyrights
 -Problems with module "resource" on Windows
 -Some small bugs
