+++ flaskJSONRPCServer/example/parallelExecuting.py
@@ -151,12 +151,12 @@ if __name__=='__main__':
    #    <tweakDescriptors> set descriptor's limit for server
    #    <jsonBackend>      set JSON backend. Auto fallback to native when problems
    #    <notifBackend>     set backend for Notify-requests
-   import ujson
    from flaskJSONRPCServer import magicDict
-   jsonBackend=magicDict({
-      'dumps': lambda data, **kwargs: ujson.dumps(data),
-      'loads': lambda data, **kwargs: ujson.loads(data)
-   })
+   # import ujson
+   # jsonBackend=magicDict({
+   #    'dumps': lambda data, **kwargs: ujson.dumps(data),
+   #    'loads': lambda data, **kwargs: ujson.loads(data)
+   # })
    server=flaskJSONRPCServer(("0.0.0.0", 7001), blocking=False, cors=True, gevent=True, debug=False, log=False, fallback=True, allowCompress=False, jsonBackend='simplejson', tweakDescriptors=[1000, 1000], dispatcherBackend='parallelWithSocket', notifBackend='simple', experimental=True)
    # Register dispatchers for single functions
    server.registerFunction(stats, path='/api', dispatcherBackend='simple')
